Setup basic Gem structure.
Create the gem using `bundle gem --test` to get all of the basic
structure for free. Switch to the Apache license since we use it for
many of our current OSS.

Targeting Ruby 2.1+ for this to keep with the newer releases. Likewise,
targeting RSpec ~> 3.2. RSpec is trying to follow strict SemVer so any
project currently on RSpec 3.0 or 3.1 should not have issues moving to
3.2. The rack-test version is used by the latest Rails and Sinatra
versions.