Configure RSpec
This adds `rspec` to the binstubs for the project. It also updates the
configuration based on the files generated by `bin/rspec --init`.

The document formatter is removed from `.rspec` so that the default
progress formatter will be used during local development. Individual
developers can choose to override this using a custom `.rspec-local`
file. Lastly, the document formatter is forced if the environment is CI
(such as Travis CI).

In the `spec/spec_helper.rb` all of the default configurations are
accepted. Additionally the older `should` syntax has been explicitly
disabled. This project will use the newer `expect` syntax and disabling
`should` helps make this explicit.

The profiler output, while very useful, is often too verbose for daily
development. This adjusts the config so that the profiler is only shown,
by default, when multiple spec files are run. This should display the
information when the entire suite is run, but hide it while a developer
works on a single spec file.