Ensure our RSpec dependency is met on load.
We are dependent on rspec-core and rack-test. Ensure both of these
dependencies are available to us before loading our code. If they are
not, we may fail in unexpected ways.

We are not attempting to `require` these dependencies. This is because
we may be outside of the expected RSpec life cycle. We do not want to
load it if this is the case, as that may cause strange behavior for them.