Base implementation for the API formatter.
This setups up the necessary infrastructure to support the general self
documenting JSON based API request specs.

To keep things simple and straight forward several data classes are
included. This allows for the different parts of the document to easily
be updated / manipulated in Ruby. It also keeps a logical separation
between the concerns.

For now, much of the storage of the information is in the actual RSpec
example group object. The DSL stores information either in the metadata
or in a class level ivar. The class ivars are not inherited by child
classes. This allows for settings to stay at the group level (i.e. an
`explanation`). The metadata is used for other settings (i.e. the
`resource`) so that it can be inherited and used by nested / child
groups. Using the metadata also means a child group can change it
without affecting any of the parent groups.

A sample template is included.

The DSL is expanded to provide accessors for storing a `request` and
`response` per example group. The intent is to use the `record` DSL
method to pick a _single_ spec per `document` group which should be
responsible for representing the URI path, headers, and body content
sent between the client and server.

This decision was made instead of attempting to have the `document`
method auto create a spec for this purpose:

- it's very difficult to create a generic automatic spec
- the various http methods (i.e. `get`, `post`) are not available in a
  group context
- it makes it a little more clear the source of `request` and `response`
- the `response` and `request` stored are fully up to the spec author
  and can be changed at the end of the spec run, if necessary

This adds the appropriate request and response headers that are captured
by `record`. It utilizes a whitelist of headers to display in the
output. This keeps the output clean and minimal, focusing only on the
headers that are actually important to the API.

Custom headers can be added to the output by using the `header` example
group macro. This stores the header, and an optional static request
header value, in the group's metadata. As with other metadata based
macros, this allows all subgroups to inherit the settings by default.

The reason the static value is not used in the response is that we
really want the actual data from the response. We allow a static value
on the request to hide things like the authentication token. This also
allows us to set a stub value for the cases, also such as with the
token, where we by pass that by using a helper, and thus it's never
actually sent.

This includes the requested URL in the curl example. It additionally
adds a macro `use_host`. This sets the `default_url_options[:host]` for
the spec. Since this is actually a global value, the helper only wraps
each spec with this setting.

The `resource_endpoint` helper in the example is the fully qualified URL
starting with this custom host. This is necessary because the controller
code under test does not respect the `default_url_options` setting
configured in the spec. Instead it looks at the request and uses the
full URL specified. If no full URL is provided it uses what is
configured in the environments file, or defaults to
"http://www.example.com" if none is set. Either way, the only way for us
to pass the custom host is to use the fully qualified URL.

An "Example" section is included for those sections which have
parameters. This allows the documentation to show, and hotlink, to a
sample of how a request should be formatted.

Sometimes these sections need additional annotation to help clarify
intent, parameters, usage, etc. This adds an `example_explanation` macro
to allow an author to include this.

This is the first pass at allowing custom sections to be included from
other contexts. Right now this is simply an alias for `include_context`.
The advantage of this is it's just RSpec. The disadvantage is it doesn't
hook cleanly into our existing document structure.

This also properly setups up line returns for the document / section
explanations. However, the logic for this is being duplicated in the
macros and the `Explainable` module. This is an unfortunate
repercussion for the initial decision to wait until the formatter runs
to build up the document structure.

We could pull the raw request json which would be a single line string.
However, for large JSON data this is difficult to read. Instead we just
wrap the multi-line JSON in single quotes which cURL can interpret.

It is valid for JSON to have newlines in it, so that is not an issue.

cURL does not like having a single quote inside of a string. The only
real way to handle it is to convert it to the UTF-8 equivalent:
`\u0027`.

Since we will be doing this for the data we might as well do it for the
headers. It is possible for someone to put a single quote in a header in
the future. This also means we no longer need to call `inspect` on the
header since we are wrapping it in single quotes.

When generating the template we define our own root. This was a quick
hack and probably needs to be updated in the future. Though it is fairly
likely the working director is the project root. In many apps, in order
to run RSpec you need to be in the project root. Additionally, Rake
tasks often need to be run from the project root as well.

ERB is a bit backwards requiring us to provide it an binding for it to
evaluate in. `binding` is defined in `Kernel` and is based on the
calling scope. So calling `obj.binding` will not give us the binding for
`obj` but instead for our caller scope. In order to properly access an
object's binding we need to be in that object. We could simply
`instance_exec` then call `binding`, but it's easier to expose the
binding.