Include rake task to run API docs.
This setups up a rake task for generating the API docs. It wraps up the
common default settings someone would want when generating the docs:

- declare the custom API formatter

- declare a 2nd formatter, as "documentation", so that the specs and
  their pass / fail status can be seen

- sets the order to defined, so that the output for each spec file is
  consistent across each generation; this also means the order that
  appears is defined by the documentation/spec author

- only run specs which should be related to the api docs; this is by
  specifying the tag `generate_doc` which is set by the `document` DSL
  method

By default running the rake task will use the development environment.
This checks to see if we have an available prepare task to use, such as
the one provided by rspec-rails. Otherwise, we create our own so we can
set the test environment.