Remove dependency on custom response monkey patch.
This removes the dependency on a custom `response` monkey patch for a
`json_body`. This patch was put in place in the host app which this gem
was extracted.

There is no need for us to monkey patch a `response` object so instead
we simply convert it into a helper method.